+++ examples/hello-world-isr/app.tsx
@@ -1,6 +1,6 @@
-import React, { ComponentType } from 'react'
+import React, { FC } from 'react'
 
-export default function App({ Page, pageProps }: { Page: ComponentType<any>, pageProps: any }) {
+export default function App({ Page, pageProps }: { Page: FC, pageProps: any }) {
   return (
     <main>
       <head>
